Move fixture to shared module
